requirements.txt
boto3==1.42.25
requests==2.32.5
openai==2.15.0
pandas==2.3.3
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
slowapi==0.1.9
pytest==9.0.2
httpx==0.28.1